Stage 1 - split into separate libraries internal to package
