+++ Sources/HummingbirdWSCompression/Exports.swift
@@ -12,4 +12,4 @@
 //
 //===----------------------------------------------------------------------===//
 
-@_exported import WSClient
+@_exported import WSCompression
