+++ Sources/WSClient/WebSocketClientError.swift
