+++ Sources/WSCore/WebSocketExtension.swift
