+++ Sources/WSCore/WebSocketFrameSequence.swift
