+++ Sources/WSCore/WebSocketMessage.swift
